requirements.txt
feedparser>=6.0.11
openai>=1.30.0
python-dotenv>=1.0.1
requests>=2.31.0
